requirements/base.txt
Django==1.6.5
bpython==0.12
django-braces==1.4.0
django-model-utils==1.5.0
django-allauth==0.17.0
django-authtools==0.2.1
logutils==0.3.3
South==1.0
django-reversion==1.8.1
Pillow==2.4.0
django-contrib-comments==1.5
django-threadedcomments==0.9.0
django-mptt==0.6.1
